requirements.txt
# === CORE AIRFLOW ===
apache-airflow==2.8.0
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-clickhouse==1.0.2
apache-airflow-providers-cassandra==3.6.0
apache-airflow-providers-celery==3.5.0
apache-airflow-providers-redis==3.5.0
apache-airflow-providers-http==4.7.0

# === DATABASE DRIVERS ===
psycopg2-binary==2.9.9
clickhouse-driver==0.4.6
cassandra-driver==3.29.1
sqlalchemy==2.0.23
pymysql==1.1.0

# === DATA PROCESSING ===
pandas==2.1.3
numpy==1.26.3
pyarrow==14.0.1
openpyxl==3.11.0

# === UTILITIES ===
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1
pytz==2023.3
python-dateutil==2.8.2

# === LOGGING & MONITORING ===
python-json-logger==2.0.7
prometheus-client==0.19.0

# === TESTING & DEVELOPMENT ===
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
isort==5.13.2

# === APACHE SPARK (опционально для HDFS) ===
pyspark==3.5.0

# === SECURITY ===
cryptography==41.0.7
